+++ src/com/nixinova/player/Player.java
@@ -1,5 +1,6 @@
 package com.nixinova.player;
 
+import com.nixinova.blocks.HoveredBlock;
 import com.nixinova.coords.BlockCoord;
 import com.nixinova.coords.Coord3;
 import com.nixinova.world.World;
@@ -8,11 +9,11 @@ public class Player {
 	public static final double PLAYER_HEIGHT = 1.8;
 
 	private Coord3 position;
-	private BlockCoord lookingAtBlock;
+	private HoveredBlock lookingAtBlock;
 
 	public Player() {
 		this.position = new Coord3();
-		this.lookingAtBlock = null;
+		this.lookingAtBlock = new HoveredBlock(null, null);
 	}
 
 	public Coord3 getPosition() {
@@ -23,18 +24,12 @@ public class Player {
 		this.position = footPosition;
 	}
 
-	public BlockCoord getLookingAt() {
+	public HoveredBlock getLookingAt() {
 		return this.lookingAtBlock;
 	}
 
-	public void setLookingAt(BlockCoord coord) {
-		this.lookingAtBlock = coord;
-	}
-
-	public void setLookingAt(int x, int y, int z) {
-		this.lookingAtBlock.x = x;
-		this.lookingAtBlock.y = y;
-		this.lookingAtBlock.z = z;
+	public void setLookingAt(HoveredBlock result) {
+		this.lookingAtBlock = result;
 	}
 
 	public boolean isWithinWorld(World world) {